requirements.txt
Django>=5.0.1
pillow>=10.2.0



channels>=4.0.0